Enables manual release triggering via workflow dispatch
Allows triggering a new release manually via "workflow_dispatch".

This change adds a workflow_dispatch trigger to the release workflow
with an input to force release creation, even if the version in
package.json hasn't changed, or the tag already exists.

It also adds logic to delete the existing tag and release before
creating a new one during manual triggering.